+++ src/main/java/net/minecraft/block/BlockCauldron.java
@@ -31,7 +31,7 @@ public class BlockCauldron extends Block {
 
     public BlockCauldron() {
         super(Material.iron, MapColor.stoneColor);
-        this.setDefaultState(this.blockState.getBaseState().withProperty(LEVEL, Integer.valueOf(0)));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(LEVEL, 0));
     }
 
     public void addCollisionBoxesToList(World worldIn, BlockPos pos, IBlockState state, AxisAlignedBB mask, List<AxisAlignedBB> list, Entity collidingEntity) {
@@ -62,7 +62,7 @@ public class BlockCauldron extends Block {
     }
 
     public void onEntityCollidedWithBlock(World worldIn, BlockPos pos, IBlockState state, Entity entityIn) {
-        int i = state.getValue(LEVEL).intValue();
+        int i = state.getValue(LEVEL);
         float f = (float) pos.getY() + (6.0F + (float) (3 * i)) / 16.0F;
 
         if (!worldIn.isRemote && entityIn.isBurning() && i > 0 && entityIn.getEntityBoundingBox().minY <= (double) f) {
@@ -80,7 +80,7 @@ public class BlockCauldron extends Block {
             if (itemstack == null) {
                 return true;
             } else {
-                int i = state.getValue(LEVEL).intValue();
+                int i = state.getValue(LEVEL);
                 Item item = itemstack.getItem();
 
                 if (item == Items.water_bucket) {
@@ -164,7 +164,7 @@ public class BlockCauldron extends Block {
     }
 
     public void setWaterLevel(World worldIn, BlockPos pos, IBlockState state, int level) {
-        worldIn.setBlockState(pos, state.withProperty(LEVEL, Integer.valueOf(MathHelper.clamp_int(level, 0, 3))), 2);
+        worldIn.setBlockState(pos, state.withProperty(LEVEL, MathHelper.clamp_int(level, 0, 3)), 2);
         worldIn.updateComparatorOutputLevel(pos, this);
     }
 
@@ -172,7 +172,7 @@ public class BlockCauldron extends Block {
         if (worldIn.rand.nextInt(20) == 1) {
             IBlockState iblockstate = worldIn.getBlockState(pos);
 
-            if (iblockstate.getValue(LEVEL).intValue() < 3) {
+            if (iblockstate.getValue(LEVEL) < 3) {
                 worldIn.setBlockState(pos, iblockstate.cycleProperty(LEVEL), 2);
             }
         }
@@ -191,15 +191,15 @@ public class BlockCauldron extends Block {
     }
 
     public int getComparatorInputOverride(World worldIn, BlockPos pos) {
-        return worldIn.getBlockState(pos).getValue(LEVEL).intValue();
+        return worldIn.getBlockState(pos).getValue(LEVEL);
     }
 
     public IBlockState getStateFromMeta(int meta) {
-        return this.getDefaultState().withProperty(LEVEL, Integer.valueOf(meta));
+        return this.getDefaultState().withProperty(LEVEL, meta);
     }
 
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(LEVEL).intValue();
+        return state.getValue(LEVEL);
     }
 
     protected BlockState createBlockState() {
