+++ src/main/java/net/minecraft/block/BlockLiquid.java
@@ -48,7 +48,7 @@ public abstract class BlockLiquid extends Block {
     }
 
     protected int getLevel(IBlockAccess worldIn, BlockPos pos) {
-        return worldIn.getBlockState(pos).getBlock().getMaterial() == this.blockMaterial ? worldIn.getBlockState(pos).getValue(LEVEL).intValue() : -1;
+        return worldIn.getBlockState(pos).getBlock().getMaterial() == this.blockMaterial ? worldIn.getBlockState(pos).getValue(LEVEL) : -1;
     }
 
     protected int getEffectiveFlowDecay(IBlockAccess worldIn, BlockPos pos) {
@@ -65,7 +65,7 @@ public abstract class BlockLiquid extends Block {
     }
 
     public boolean canCollideCheck(IBlockState state, boolean hitIfLiquid) {
-        return hitIfLiquid && state.getValue(LEVEL).intValue() == 0;
+        return hitIfLiquid && state.getValue(LEVEL) == 0;
     }
 
     public boolean isBlockSolid(IBlockAccess worldIn, BlockPos pos, EnumFacing side) {
@@ -132,7 +132,7 @@ public abstract class BlockLiquid extends Block {
             }
         }
 
-        if (worldIn.getBlockState(pos).getValue(LEVEL).intValue() >= 8) {
+        if (worldIn.getBlockState(pos).getValue(LEVEL) >= 8) {
             for (EnumFacing enumfacing1 : EnumFacing.Plane.HORIZONTAL) {
                 BlockPos blockpos1 = pos.offset(enumfacing1);
 
@@ -174,7 +174,7 @@ public abstract class BlockLiquid extends Block {
         double d2 = pos.getZ();
 
         if (this.blockMaterial == Material.water) {
-            int i = state.getValue(LEVEL).intValue();
+            int i = state.getValue(LEVEL);
 
             if (i > 0 && i < 8) {
                 if (rand.nextInt(64) == 0) {
@@ -243,13 +243,13 @@ public abstract class BlockLiquid extends Block {
             if (flag) {
                 Integer integer = state.getValue(LEVEL);
 
-                if (integer.intValue() == 0) {
+                if (integer == 0) {
                     worldIn.setBlockState(pos, Blocks.obsidian.getDefaultState());
                     this.triggerMixEffects(worldIn, pos);
                     return true;
                 }
 
-                if (integer.intValue() <= 4) {
+                if (integer <= 4) {
                     worldIn.setBlockState(pos, Blocks.cobblestone.getDefaultState());
                     this.triggerMixEffects(worldIn, pos);
                     return true;
@@ -276,7 +276,7 @@ public abstract class BlockLiquid extends Block {
     }
 
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(LEVEL).intValue();
+        return state.getValue(LEVEL);
     }
 
     protected BlockState createBlockState() {
