+++ src/main/java/net/minecraft/block/BlockPressurePlateWeighted.java
@@ -37,7 +37,7 @@ public class BlockPressurePlateWeighted extends BlockBasePressurePlate {
     }
 
     protected int getRedstoneStrength(IBlockState state) {
-        return state.getValue(POWER).intValue();
+        return state.getValue(POWER);
     }
 
     protected IBlockState setRedstoneStrength(IBlockState state, int strength) {
@@ -53,7 +53,7 @@ public class BlockPressurePlateWeighted extends BlockBasePressurePlate {
     }
 
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(POWER).intValue();
+        return state.getValue(POWER);
     }
 
     protected BlockState createBlockState() {
