+++ src/main/java/net/minecraft/block/BlockStem.java
@@ -29,7 +29,7 @@ public class BlockStem extends BlockBush implements IGrowable {
     private final Block crop;
 
     protected BlockStem(Block crop) {
-        this.setDefaultState(this.blockState.getBaseState().withProperty(AGE, Integer.valueOf(0)).withProperty(FACING, EnumFacing.UP));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(AGE, 0).withProperty(FACING, EnumFacing.UP));
         this.crop = crop;
         this.setTickRandomly(true);
         float f = 0.125F;
@@ -61,10 +61,10 @@ public class BlockStem extends BlockBush implements IGrowable {
             float f = BlockCrops.getGrowthChance(this, worldIn, pos);
 
             if (rand.nextInt((int) (25.0F / f) + 1) == 0) {
-                int i = state.getValue(AGE).intValue();
+                int i = state.getValue(AGE);
 
                 if (i < 7) {
-                    state = state.withProperty(AGE, Integer.valueOf(i + 1));
+                    state = state.withProperty(AGE, i + 1);
                     worldIn.setBlockState(pos, state, 2);
                 } else {
                     for (EnumFacing enumfacing : EnumFacing.Plane.HORIZONTAL) {
@@ -85,15 +85,15 @@ public class BlockStem extends BlockBush implements IGrowable {
     }
 
     public void growStem(World worldIn, BlockPos pos, IBlockState state) {
-        int i = state.getValue(AGE).intValue() + MathHelper.getRandomIntegerInRange(worldIn.rand, 2, 5);
-        worldIn.setBlockState(pos, state.withProperty(AGE, Integer.valueOf(Math.min(7, i))), 2);
+        int i = state.getValue(AGE) + MathHelper.getRandomIntegerInRange(worldIn.rand, 2, 5);
+        worldIn.setBlockState(pos, state.withProperty(AGE, Math.min(7, i)), 2);
     }
 
     public int getRenderColor(IBlockState state) {
         if (state.getBlock() != this) {
             return super.getRenderColor(state);
         } else {
-            int i = state.getValue(AGE).intValue();
+            int i = state.getValue(AGE);
             int j = i * 32;
             int k = 255 - i * 8;
             int l = i * 4;
@@ -111,7 +111,7 @@ public class BlockStem extends BlockBush implements IGrowable {
     }
 
     public void setBlockBoundsBasedOnState(IBlockAccess worldIn, BlockPos pos) {
-        this.maxY = (float) (worldIn.getBlockState(pos).getValue(AGE).intValue() * 2 + 2) / 16.0F;
+        this.maxY = (float) (worldIn.getBlockState(pos).getValue(AGE) * 2 + 2) / 16.0F;
         float f = 0.125F;
         this.setBlockBounds(0.5F - f, 0.0F, 0.5F - f, 0.5F + f, (float) this.maxY, 0.5F + f);
     }
@@ -123,7 +123,7 @@ public class BlockStem extends BlockBush implements IGrowable {
             Item item = this.getSeedItem();
 
             if (item != null) {
-                int i = state.getValue(AGE).intValue();
+                int i = state.getValue(AGE);
 
                 for (int j = 0; j < 3; ++j) {
                     if (worldIn.rand.nextInt(15) <= i) {
@@ -148,7 +148,7 @@ public class BlockStem extends BlockBush implements IGrowable {
     }
 
     public boolean canGrow(World worldIn, BlockPos pos, IBlockState state, boolean isClient) {
-        return state.getValue(AGE).intValue() != 7;
+        return state.getValue(AGE) != 7;
     }
 
     public boolean canUseBonemeal(World worldIn, Random rand, BlockPos pos, IBlockState state) {
@@ -160,11 +160,11 @@ public class BlockStem extends BlockBush implements IGrowable {
     }
 
     public IBlockState getStateFromMeta(int meta) {
-        return this.getDefaultState().withProperty(AGE, Integer.valueOf(meta));
+        return this.getDefaultState().withProperty(AGE, meta);
     }
 
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(AGE).intValue();
+        return state.getValue(AGE);
     }
 
     protected BlockState createBlockState() {
