+++ src/main/java/net/minecraft/client/audio/SoundManager.java
@@ -184,13 +184,13 @@ public class SoundManager {
             ISound isound = entry.getValue();
 
             if (!this.sndSystem.playing(s1)) {
-                int i = this.playingSoundsStopTime.get(s1).intValue();
+                int i = this.playingSoundsStopTime.get(s1);
 
                 if (i <= this.playTime) {
                     int j = isound.getRepeatDelay();
 
                     if (isound.canRepeat() && j > 0) {
-                        this.delayedSounds.put(isound, Integer.valueOf(this.playTime + j));
+                        this.delayedSounds.put(isound, this.playTime + j);
                     }
 
                     iterator.remove();
@@ -217,7 +217,7 @@ public class SoundManager {
         while (iterator1.hasNext()) {
             Entry<ISound, Integer> entry1 = iterator1.next();
 
-            if (this.playTime >= entry1.getValue().intValue()) {
+            if (this.playTime >= entry1.getValue()) {
                 ISound isound1 = entry1.getKey();
 
                 if (isound1 instanceof ITickableSound) {
@@ -235,7 +235,7 @@ public class SoundManager {
             return false;
         } else {
             String s = this.invPlayingSounds.get(sound);
-            return s == null ? false : this.sndSystem.playing(s) || this.playingSoundsStopTime.containsKey(s) && this.playingSoundsStopTime.get(s).intValue() <= this.playTime;
+            return s == null ? false : this.sndSystem.playing(s) || this.playingSoundsStopTime.containsKey(s) && this.playingSoundsStopTime.get(s) <= this.playTime;
         }
     }
 
@@ -292,7 +292,7 @@ public class SoundManager {
                             this.sndSystem.setPitch(s, (float) d0);
                             this.sndSystem.setVolume(s, f2);
                             this.sndSystem.play(s);
-                            this.playingSoundsStopTime.put(s, Integer.valueOf(this.playTime + 20));
+                            this.playingSoundsStopTime.put(s, this.playTime + 20);
                             this.playingSounds.put(s, p_sound);
                             this.playingSoundPoolEntries.put(p_sound, soundpoolentry);
 
@@ -333,7 +333,7 @@ public class SoundManager {
     }
 
     public void playDelayedSound(ISound sound, int delay) {
-        this.delayedSounds.put(sound, Integer.valueOf(this.playTime + delay));
+        this.delayedSounds.put(sound, this.playTime + delay);
     }
 
     private static URL getURLForSoundResource(final ResourceLocation p_148612_0_) {
