+++ src/main/java/net/minecraft/command/PlayerSelector.java
@@ -263,11 +263,11 @@ public class PlayerSelector {
                         Score score = scoreboard.getValueFromObjective(s1, scoreobjective);
                         int i = score.getScorePoints();
 
-                        if (i < entry.getValue().intValue() && flag) {
+                        if (i < entry.getValue() && flag) {
                             return false;
                         }
 
-                        if (i > entry.getValue().intValue() && !flag) {
+                        if (i > entry.getValue() && !flag) {
                             return false;
                         }
                     }
@@ -493,7 +493,7 @@ public class PlayerSelector {
 
         for (String s : p_96560_0_.keySet()) {
             if (s.startsWith("score_") && s.length() > "score_".length()) {
-                map.put(s.substring("score_".length()), Integer.valueOf(MathHelper.parseIntWithDefault(p_96560_0_.get(s), 1)));
+                map.put(s.substring("score_".length()), MathHelper.parseIntWithDefault(p_96560_0_.get(s), 1));
             }
         }
