+++ src/main/java/net/minecraft/entity/monster/EntityGhast.java
@@ -42,7 +42,7 @@ public class EntityGhast extends EntityFlying implements IMob {
     }
 
     public void setAttacking(boolean attacking) {
-        this.dataWatcher.updateObject(16, Byte.valueOf((byte) (attacking ? 1 : 0)));
+        this.dataWatcher.updateObject(16, (byte) (attacking ? 1 : 0));
     }
 
     public int getFireballStrength() {
@@ -71,7 +71,7 @@ public class EntityGhast extends EntityFlying implements IMob {
 
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(16, Byte.valueOf((byte) 0));
+        this.dataWatcher.addObject(16, (byte) 0);
     }
 
     protected void applyEntityAttributes() {
