+++ src/main/java/net/minecraft/entity/passive/EntityOcelot.java
@@ -53,7 +53,7 @@ public class EntityOcelot extends EntityTameable {
 
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(18, Byte.valueOf((byte) 0));
+        this.dataWatcher.addObject(18, (byte) 0);
     }
 
     public void updateAITasks() {
@@ -205,7 +205,7 @@ public class EntityOcelot extends EntityTameable {
     }
 
     public void setTameSkin(int skinId) {
-        this.dataWatcher.updateObject(18, Byte.valueOf((byte) skinId));
+        this.dataWatcher.updateObject(18, (byte) skinId);
     }
 
     public boolean getCanSpawnHere() {
